doc(uds-2-np): Add todo for handshake in process_pipe_connection
Seems that sending the cookie might not be enough to establish the
connection, as things get stuck after this with no
messages one way or the other. Calling out that we need to investigate
how to complete any handshake for libassuan's AF_UNIX emulation before
we can continue communicating between the pipe and socket.